pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quorum-core"
version = "0.1.0"
description = "Consensus engine for distributed agent systems"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
  "mypy>=1.11",
  "build>=1.2",
]
quantum = [
  "qiskit>=1.1",
  "qiskit-aer>=0.15",
  "qiskit-optimization>=0.6",
  "dwave-ocean-sdk>=9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.11"
files = ["src", "tests"]
ignore_missing_imports = true
warn_unused_ignores = true
show_error_codes = true
